backend/pyproject.toml
[project]
name = "network-automation-backend"
version = "0.1.0"
description = "Network automation web application backend"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "Django>=5.1",
    "djangorestframework>=3.15.2",
    "djangorestframework-simplejwt>=5.3.1",
    "channels>=4.1.0",
    "channels-redis>=4.2.0",
    "daphne>=4.1.2",
    "celery>=5.3.6",
    "redis>=5.0.1",
    "psycopg2-binary>=2.9.9",
    "django-filter>=24.3",
    "whitenoise>=6.6.0",
    "argon2-cffi>=23.1.0",
    "prometheus-client>=0.19.0",
    "cryptography>=43.0.0",
    "pyyaml>=6.0.1",
    "jinja2>=3.1.3",
    "python-multipart>=0.0.6",
    "napalm>=4.1.0",
    "netmiko>=4.4.0",
    "nornir>=3.4.1",
    "nornir-napalm>=0.4.0",
    "nornir-netmiko>=1.0.0",
    "asyncssh>=2.14.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-django>=4.8.0",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "mypy>=1.8.0",
    "ruff>=0.1.14",
    "black>=24.1.1",
    "django-stubs>=4.2.7",
    "djangorestframework-stubs>=3.15.0",
    "types-redis>=4.6.0.20241004",
    "types-requests>=2.32.0.20241016",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = [
    "webnet",
    "webnet.core",
    "webnet.users",
    "webnet.customers",
    "webnet.devices",
    "webnet.jobs",
    "webnet.config_mgmt",
    "webnet.compliance",
    "webnet.networkops",
    "webnet.api",
    "webnet.ui",
]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
exclude = "migrations"

